+++ packages/editor/api-report.md
@@ -17,8 +17,18 @@ import { EASINGS } from '@tldraw/primitives';
 import { EmbedDefinition } from '@tldraw/tlschema';
 import { EventEmitter } from 'eventemitter3';
 import { getHashForString } from '@tldraw/utils';
+import { getIndexAbove } from '@tldraw/indices';
+import { getIndexBelow } from '@tldraw/indices';
+import { getIndexBetween } from '@tldraw/indices';
+import { getIndexGenerator } from '@tldraw/indices';
+import { getIndices } from '@tldraw/indices';
+import { getIndicesAbove } from '@tldraw/indices';
+import { getIndicesBelow } from '@tldraw/indices';
+import { getIndicesBetween } from '@tldraw/indices';
+import { getMaxIndex } from '@tldraw/indices';
 import { HistoryEntry } from '@tldraw/tlstore';
 import { ID } from '@tldraw/tlstore';
+import { indexGenerator } from '@tldraw/indices';
 import { MatLike } from '@tldraw/primitives';
 import { Matrix2d } from '@tldraw/primitives';
 import { Matrix2dModel } from '@tldraw/primitives';
@@ -33,6 +43,8 @@ import { SelectionEdge } from '@tldraw/primitives';
 import { SelectionHandle } from '@tldraw/primitives';
 import { SerializedSchema } from '@tldraw/tlstore';
 import { Signal } from 'signia';
+import { sortById } from '@tldraw/indices';
+import { sortByIndex } from '@tldraw/indices';
 import { StoreSchema } from '@tldraw/tlstore';
 import { StoreSnapshot } from '@tldraw/tlstore';
 import { StoreValidator } from '@tldraw/tlstore';
@@ -754,32 +766,23 @@ export function getImageSizeFromSrc(dataURL: string): Promise<{
 // @public
 export function getIncrementedName(name: string, others: string[]): string;
 
-// @public (undocumented)
-export function getIndexAbove(below: string): string;
+export { getIndexAbove }
 
-// @public (undocumented)
-export function getIndexBelow(above: string): string;
+export { getIndexBelow }
 
-// @public (undocumented)
-export function getIndexBetween(below: string, above?: string): string;
+export { getIndexBetween }
 
-// @public (undocumented)
-export function getIndexGenerator(): () => string;
+export { getIndexGenerator }
 
-// @public (undocumented)
-export function getIndices(n: number): string[];
+export { getIndices }
 
-// @public (undocumented)
-export function getIndicesAbove(below: string, n: number): string[];
+export { getIndicesAbove }
 
-// @public (undocumented)
-export function getIndicesBelow(above: string, n: number): string[];
+export { getIndicesBelow }
 
-// @public (undocumented)
-export function getIndicesBetween(below: string | undefined, above: string | undefined, n: number): string[];
+export { getIndicesBetween }
 
-// @public (undocumented)
-export function getMaxIndex(...indices: (string | undefined)[]): string;
+export { getMaxIndex }
 
 // @public
 export function getMediaAssetFromFile(file: File): Promise<TLAsset>;
@@ -883,8 +886,7 @@ export const ICON_SIZES: Record<TLSizeType, number>;
 // @public (undocumented)
 export const INDENT = "  ";
 
-// @public (undocumented)
-export function indexGenerator(n?: number): Generator<string, void, unknown>;
+export { indexGenerator }
 
 // @public (undocumented)
 export interface InitializingSyncedStore {
@@ -1470,15 +1472,9 @@ export function setRuntimeOverrides(input: Partial<typeof runtime>): void;
 // @public (undocumented)
 export function snapToGrid(n: number, gridSize: number): number;
 
-// @public (undocumented)
-export function sortById<T extends {
-    id: string;
-}>(a: T, b: T): -1 | 0 | 1;
+export { sortById }
 
-// @public (undocumented)
-export function sortByIndex<T extends {
-    index: string;
-}>(a: T, b: T): -1 | 0 | 1;
+export { sortByIndex }
 
 // @public (undocumented)
 export abstract class StateNode implements Partial<TLEventHandlers> {
