+++ packages/editor/src/lib/app/App.ts
@@ -1,24 +1,29 @@
 import {
-	approximately,
-	areAnglesCompatible,
+	getIndexAbove,
+	getIndexBetween,
+	getIndices,
+	getIndicesAbove,
+	getIndicesBetween,
+	sortById,
+	sortByIndex,
+} from '@tldraw/indices'
+import {
 	Box2d,
-	clamp,
 	EASINGS,
-	intersectPolygonPolygon,
 	MatLike,
 	Matrix2d,
 	Matrix2dModel,
 	PI2,
-	pointInPolygon,
 	Vec2d,
 	VecLike,
+	approximately,
+	areAnglesCompatible,
+	clamp,
+	intersectPolygonPolygon,
+	pointInPolygon,
 } from '@tldraw/primitives'
 import {
 	Box2dModel,
-	createCustomShapeId,
-	createShapeId,
-	isShape,
-	isShapeId,
 	TLArrowShape,
 	TLAsset,
 	TLAssetId,
@@ -54,14 +59,18 @@ import {
 	TLUserId,
 	TLVideoAsset,
 	Vec2dModel,
+	createCustomShapeId,
+	createShapeId,
+	isShape,
+	isShapeId,
 } from '@tldraw/tlschema'
 import { BaseRecord, ComputedCache, HistoryEntry } from '@tldraw/tlstore'
 import { annotateError, compact, dedupe, deepCopy, partition, structuredClone } from '@tldraw/utils'
 import { EventEmitter } from 'eventemitter3'
 import { nanoid } from 'nanoid'
-import { atom, computed, EMPTY_ARRAY, transact } from 'signia'
-import { TldrawEditorConfig } from '../config/TldrawEditorConfig'
+import { EMPTY_ARRAY, atom, computed, transact } from 'signia'
 import { TLShapeDef } from '../config/TLShapeDefinition'
+import { TldrawEditorConfig } from '../config/TldrawEditorConfig'
 import {
 	ANIMATION_MEDIUM_MS,
 	BLACKLISTED_PROPS,
@@ -79,27 +88,18 @@ import {
 	MAX_PAGES,
 	MAX_SHAPES_PER_PAGE,
 	MAX_ZOOM,
-	MIN_ZOOM,
 	MINOR_NUDGE_FACTOR,
+	MIN_ZOOM,
 	STYLES,
 	SVG_PADDING,
 	ZOOMS,
 } from '../constants'
 import { exportPatternSvgDefs } from '../hooks/usePattern'
+import { WeakMapCache } from '../utils/WeakMapCache'
 import { dataUrlToFile, getMediaAssetFromFile } from '../utils/assets'
 import { getIncrementedName, uniqueId } from '../utils/data'
 import { setPropsForNextShape } from '../utils/props-for-next-shape'
-import {
-	getIndexAbove,
-	getIndexBetween,
-	getIndices,
-	getIndicesAbove,
-	getIndicesBetween,
-	sortById,
-	sortByIndex,
-} from '../utils/reordering/reordering'
 import { applyRotationToSnapshotShapes, getRotationSnapshot } from '../utils/rotation'
-import { WeakMapCache } from '../utils/WeakMapCache'
 import { arrowBindingsIndex } from './derivations/arrowBindingsIndex'
 import { parentsToChildrenWithIndexes } from './derivations/parentsToChildrenWithIndexes'
 import { shapeIdsInCurrentPage } from './derivations/shapeIdsInCurrentPage'
@@ -111,18 +111,18 @@ import { HistoryManager } from './managers/HistoryManager'
 import { SnapManager } from './managers/SnapManager'
 import { TextManager } from './managers/TextManager'
 import { TickManager } from './managers/TickManager'
-import { TLExportColors } from './shapeutils/shared/TLExportColors'
+import { TLArrowShapeDef } from './shapeutils/TLArrowUtil/TLArrowUtil'
 import { getCurvedArrowInfo } from './shapeutils/TLArrowUtil/arrow/curved-arrow'
 import {
 	getArrowTerminalsInArrowSpace,
 	getIsArrowStraight,
 } from './shapeutils/TLArrowUtil/arrow/shared'
 import { getStraightArrowInfo } from './shapeutils/TLArrowUtil/arrow/straight-arrow'
-import { TLArrowShapeDef } from './shapeutils/TLArrowUtil/TLArrowUtil'
 import { TLFrameShapeDef } from './shapeutils/TLFrameUtil/TLFrameUtil'
 import { TLGroupShapeDef } from './shapeutils/TLGroupUtil/TLGroupUtil'
 import { TLResizeMode, TLShapeUtil } from './shapeutils/TLShapeUtil'
 import { TLTextShapeDef } from './shapeutils/TLTextUtil/TLTextUtil'
+import { TLExportColors } from './shapeutils/shared/TLExportColors'
 import { RootState } from './statechart/RootState'
 import { StateNode } from './statechart/StateNode'
 import { TLClipboardModel } from './types/clipboard-types'
