+++ packages/editor/src/lib/app/derivations/parentsToChildrenWithIndexes.test.ts
@@ -1,6 +1,6 @@
+import { getIndexAbove, getIndexBetween } from '@tldraw/indices'
 import { createCustomShapeId } from '@tldraw/tlschema'
 import { TestApp } from '../../test/TestApp'
-import { getIndexAbove, getIndexBetween } from '../../utils/reordering/reordering'
 
 let app: TestApp
