+++ packages/editor/src/lib/app/statechart/TLLineTool/children/Pointing.ts
@@ -1,7 +1,7 @@
+import { getIndexAbove, sortByIndex } from '@tldraw/indices'
 import { Matrix2d, Vec2d } from '@tldraw/primitives'
 import { TLHandle, TLLineShape, TLShapeId, TLShapeType, createShapeId } from '@tldraw/tlschema'
 import { last, structuredClone } from '@tldraw/utils'
-import { getIndexAbove, sortByIndex } from '../../../../utils/reordering/reordering'
 import { TLEventHandlers, TLInterruptEvent } from '../../../types/event-types'
 import { StateNode } from '../../StateNode'
 import { TLLineTool } from '../TLLineTool'
