+++ packages/indices/api-report.md
@@ -1,0 +1,49 @@
+## API Report File for "@tldraw/indices"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+// @public
+export function getIndexAbove(below: string): string;
+
+// @public
+export function getIndexBelow(above: string): string;
+
+// @public
+export function getIndexBetween(below: string, above?: string): string;
+
+// @public
+export function getIndexGenerator(): () => string;
+
+// @public
+export function getIndices(n: number, start?: string): string[];
+
+// @public
+export function getIndicesAbove(below: string, n: number): string[];
+
+// @public
+export function getIndicesBelow(above: string, n: number): string[];
+
+// @public
+export function getIndicesBetween(below: string | undefined, above: string | undefined, n: number): string[];
+
+// @public
+export function getMaxIndex(...indices: (string | undefined)[]): string;
+
+// @public
+export function indexGenerator(n?: number): Generator<string, void, unknown>;
+
+// @public
+export function sortById<T extends {
+    id: string;
+}>(a: T, b: T): -1 | 0 | 1;
+
+// @public
+export function sortByIndex<T extends {
+    index: string;
+}>(a: T, b: T): -1 | 0 | 1;
+
+// (No @packageDocumentation comment for this package)
+
+```
