+++ public-yarn.lock
@@ -4610,6 +4610,7 @@ __metadata:
     "@peculiar/webcrypto": ^1.4.0
     "@testing-library/jest-dom": ^5.16.5
     "@testing-library/react": ^14.0.0
+    "@tldraw/indices": "workspace:*"
     "@tldraw/primitives": "workspace:*"
     "@tldraw/tlschema": "workspace:*"
     "@tldraw/tlstore": "workspace:*"
@@ -4655,6 +4656,15 @@ __metadata:
     "@tldraw/tlvalidate": "workspace:*"
     "@tldraw/ui": "workspace:*"
     "@tldraw/utils": "workspace:*"
+    lazyrepo: 0.0.0-alpha.26
+  languageName: unknown
+  linkType: soft
+
+"@tldraw/indices@workspace:*, @tldraw/indices@workspace:packages/indices":
+  version: 0.0.0-use.local
+  resolution: "@tldraw/indices@workspace:packages/indices"
+  dependencies:
+    benchmark: ^2.1.4
     lazyrepo: 0.0.0-alpha.26
   languageName: unknown
   linkType: soft
